Move posts' form-related features to form directory under posts/components, to match structure used by replies - Move AttachmentInput, PostForm to src/posts/components/form - Update CreatePost, EditPost's imports
