+++ frontend/src/features/posts/components/form/AttachmentInput.jsx
@@ -1,6 +1,6 @@
 import { useEffect, useState, useContext } from "react";
-import { PostFormContext } from "../context/PostFormContext";
-import StatusNotification from "../../../components/status-update/StatusNotification";
+import { PostFormContext } from "../../context/PostFormContext";
+import StatusNotification from "../../../../components/status-update/StatusNotification";
 
 const AttachmentInput = ({ imgCdn, serverStatus }) => {
   const { post, dispatch } = useContext(PostFormContext);
