+++ frontend/src/pages/CreatePost.jsx
@@ -1,4 +1,4 @@
-import PostForm from "../features/posts/components/PostForm";
+import PostForm from "../features/posts/components/form/PostForm";
 import { PostFormContextProvider } from "../features/posts/context/PostFormContext";
 import { useOutletContext } from "react-router-dom";
 const CreatePost = () => {
