+++ frontend/src/pages/EditPost.jsx
@@ -2,7 +2,7 @@ import { useParams } from "react-router-dom";
 import { useEffect, useState } from "react";
 import { useOutletContext } from "react-router-dom";
 import { usePost } from "../features/posts/hooks";
-import PostForm from "../features/posts/components/PostForm";
+import PostForm from "../features/posts/components/form/PostForm";
 import { PostFormContextProvider } from "../features/posts/context/PostFormContext";
 
 const EditPost = () => {
